+++ packages/mobile/src/components/Message/Message.types.ts
@@ -1,5 +1,7 @@
-import { DisplayableMessage } from '@quiet/state-manager'
+import { DisplayableMessage, DownloadStatus, FileMetadata } from '@quiet/state-manager'
 
 export interface MessageProps {
   data: DisplayableMessage[]
+  downloadStatus?: DownloadStatus
+  openImagePreview?: (medi: FileMetadata) => void
 }