+++ packages/mobile/src/store/store.ts
@@ -1,4 +1,4 @@
-import AsyncStorage from '@react-native-async-storage/async-storage'
+import FilesystemStorage from 'redux-persist-filesystem-storage'
 import { configureStore, getDefaultMiddleware } from '@reduxjs/toolkit'
 import { persistReducer, persistStore } from 'redux-persist'
 import createSagaMiddleware from 'redux-saga'
@@ -12,12 +12,13 @@ import { storeKeys as StateManagerStoreKeys, PublicChannelsTransform, MessagesTr
 const persistedReducer = persistReducer(
   {
     key: 'persistedReducer',
-    storage: AsyncStorage,
+    storage: FilesystemStorage,
     whitelist: [
       StateManagerStoreKeys.Identity,
       StateManagerStoreKeys.Communities,
       StateManagerStoreKeys.PublicChannels,
-      StateManagerStoreKeys.Messages
+      StateManagerStoreKeys.Messages,
+      StateManagerStoreKeys.Files
     ],
     transforms: [PublicChannelsTransform, MessagesTransform, FilesTransform]
   },