+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.1](https://github.com/TryQuiet/monorepo/compare/quiet@0.11.0...quiet@0.11.1) (2022-09-13)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
 # [0.11.0](https://github.com/TryQuiet/quiet/compare/quiet@0.10.2-alpha.0...quiet@0.11.0) (2022-09-07)
 
 **Note:** Version bump only for package quiet